moved json service from utils to services
